Got kernel rendering on screen!

+++ shaders/draw2D_new.vert
@@ -1,6 +1,6 @@
 #version 330 core
 
-in vec2 position;
+in ivec2 position;
 
 void main()
 {
